requirements.txt
flask==2.1.3
Flask-MySQL
pyyaml
Werkzeug==2.2.2
